Rename graphene::schema to steem::schema #1529 #1897
find libraries/schema -type f -print0 | xargs -0 sed -i -e 's/graphene/steem/'
git mv libraries/schema/include/graphene libraries/schema/include/steem
