--- libraries/schema/include/graphene/schema/schema.hpp
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <graphene/schema/abstract_schema.hpp>
-#include <graphene/schema/schema_impl.hpp>
-#include <graphene/schema/schema_types.hpp>
