--- libraries/schema/include/graphene/schema/schema_types.hpp
@@ -1,10 +1,0 @@
-
-#pragma once
-
-#include <graphene/schema/schema_types/fixed_string.hpp>
-#include <graphene/schema/schema_types/flat_map.hpp>
-#include <graphene/schema/schema_types/flat_set.hpp>
-#include <graphene/schema/schema_types/static_variant.hpp>
-#include <graphene/schema/schema_types/pair.hpp>
-#include <graphene/schema/schema_types/vector.hpp>
-
