+++ libraries/schema/include/steem/schema/schema.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <steem/schema/abstract_schema.hpp>
+#include <steem/schema/schema_impl.hpp>
+#include <steem/schema/schema_types.hpp>
