+++ libraries/schema/include/steem/schema/schema_impl.hpp
@@ -11,7 +11,7 @@
 #include <string>
 #include <vector>
 
-namespace graphene { namespace schema { namespace detail {
+namespace steem { namespace schema { namespace detail {
 
 struct get_deps_member_visitor
 {
