+++ libraries/schema/include/steem/schema/schema_types.hpp
@@ -1,0 +1,10 @@
+
+#pragma once
+
+#include <steem/schema/schema_types/fixed_string.hpp>
+#include <steem/schema/schema_types/flat_map.hpp>
+#include <steem/schema/schema_types/flat_set.hpp>
+#include <steem/schema/schema_types/static_variant.hpp>
+#include <steem/schema/schema_types/pair.hpp>
+#include <steem/schema/schema_types/vector.hpp>
+
