+++ libraries/schema/include/steem/schema/schema_types/flat_map.hpp
@@ -1,12 +1,12 @@
 
 #pragma once
 
-#include <graphene/schema/abstract_schema.hpp>
-#include <graphene/schema/schema_impl.hpp>
+#include <steem/schema/abstract_schema.hpp>
+#include <steem/schema/schema_impl.hpp>
 
 #include <boost/container/flat_map.hpp>
 
-namespace graphene { namespace schema { namespace detail {
+namespace steem { namespace schema { namespace detail {
 
 //////////////////////////////////////////////
 // flat_map                                 //
